fix(product-client): split transcript diagnostics out of the login-reachable module (PRO-187, r11)
CI's Login first-load budget gate failed at 502160 B > cap 502000 B. Root
cause: recordTranscriptPinTransition/recordTranscriptUserScrollIntent/
recordTranscriptVirtualizerBlank, added to
renderer-diagnostic-migrations.ts by the pin-transition diagnostics rung,
share that file with recordTurnEnded. use-turn-end-diagnostics.ts imports
recordTurnEnded and is mounted unconditionally by ProductLifecycleRoot
(no lazy() boundary), so the whole module -- including the transcript-only
exports nothing on /login ever calls -- travels into the main chunk the
login route loads.

Move the transcript-only records into a new sibling module,
renderer-diagnostic-migrations-transcript.ts, imported only by the
chat/transcript UI hooks (mounted inside the AuthenticatedProductClient
lazy chunk). Local measurement: index chunk gzip JS 500392 B -> 500203 B
(the pin_transition/user_scroll_intent strings move from the login chunk
into AuthenticatedProductClient-*.js). Behavior unchanged: same
recordRendererDiagnostic() calls, same call sites, just relocated;
existing r11 unit tests (renderer-diagnostic-migrations.test.ts,
use-transcript-stick-to-bottom.diagnostics.test.tsx,
use-transcript-virtualizer-blank-fallback.test.tsx) pass unmodified in
assertions.
