+++ apps/packages/product-client/src/hooks/chat/ui/transcript-reading-position-store.ts
@@ -19,7 +19,7 @@
 // "estimates cannot skew the restore" guarantee.
 
 import type { MutableRefObject, RefObject } from "react";
-import type { TranscriptPinTransitionCause } from "#product/lib/infra/diagnostics/renderer-diagnostic-migrations";
+import type { TranscriptPinTransitionCause } from "#product/lib/infra/diagnostics/renderer-diagnostic-migrations-transcript";
 
 // The working set of concurrently open sessions in one runtime is small; this
 // caps the retained set so a very long-lived tab that visits thousands of
