+++ apps/packages/product-client/src/hooks/chat/ui/use-transcript-submit-stamp-repin.ts
@@ -1,5 +1,5 @@
 import { useLayoutEffect, useRef } from "react";
-import type { TranscriptPinTransitionCause } from "#product/lib/infra/diagnostics/renderer-diagnostic-migrations";
+import type { TranscriptPinTransitionCause } from "#product/lib/infra/diagnostics/renderer-diagnostic-migrations-transcript";
 
 export interface UseTranscriptSubmitStampRepinOptions {
   /**
